imp: Move Env struct to env mod file
